transaction/requirements.txt
flask
requests
boto3
python-dotenv
pyjwt
requests
flask-cors
cryptography